Fuse mref and xref
