+++ aleph/queues.py
@@ -10,7 +10,6 @@ OP_INGEST = "ingest"
 OP_ANALYZE = "analyze"
 OP_INDEX = "index"
 OP_XREF = "xref"
-OP_XREF_ITEM = "xitem"
 OP_REINGEST = "reingest"
 OP_REINDEX = "reindex"
 OP_LOAD_MAPPING = "loadmapping"
